@@ -2,12 +2,14 @@
   "name": "docker-storm",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "./dist/client/index.jsx",
+  "type": "module",
+
   "scripts": {
     "dev": "webpack-dev-server --mode development --open --hot",
-    "start": "NODE_ENV=development webpack-dev-server --open & NODE_ENV=development nodemon server/server.js",
+    "start": "NODE_ENV=development webpack-dev-server --open & NODE_ENV=development nodemon dist/server/server.js",
     "build": "NODE_ENV=production webpack",
-    "server": "NODE_ENV=production nodemon server/server.js"
+    "server": "NODE_ENV=production nodemon dist/server/server.js"
   },
   "repository": {
     "type": "git",
@@ -20,6 +22,7 @@
   },
   "homepage": "https://github.com/oslabs-beta/Docker-Storm#readme",
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "express": "^4.18.2",
     "node-fetch": "^3.2.10",
@@ -32,14 +35,18 @@
     "@babel/core": "^7.19.6",
     "@babel/preset-env": "^7.19.4",
     "@babel/preset-react": "^7.18.6",
+    "@types/bcryptjs": "^2.4.2",
     "@types/cookie-parser": "^1.4.3",
     "@types/cors": "^2.8.12",
+    "@types/pg": "^8.6.5",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.9",
+    "@typescript-eslint/eslint-plugin": "^5.44.0",
+    "@typescript-eslint/parser": "^5.44.0",
     "babel-loader": "^9.0.0",
     "cors": "^2.8.5",
     "css-loader": "^6.7.1",
-    "eslint": "^8.26.0",
+    "eslint": "^8.28.0",
     "eslint-plugin-react": "^7.31.10",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
