@@ -1,16 +1,20 @@
-const { Pool } = require('pg');
+import pkg from 'pg';
 
-const PG_URI: string = 'postgres://fehkjvko:W4u1aDnKyCZaUaMZ37OyLZTP1UelhjRR@heffalump.db.elephantsql.com/fehkjvko'
+const { Pool } = pkg;
+
+const PG_URI = 'postgres://fehkjvko:W4u1aDnKyCZaUaMZ37OyLZTP1UelhjRR@heffalump.db.elephantsql.com/fehkjvko';
 
 
 const pool = new Pool({
-    connectionString: PG_URI,
-  });
+  connectionString: PG_URI,
+});
+
+const db = {  
+  query: (text: string, params: string[]) => {
+    console.log('executed query', text);
+    return pool.query(text, params);
+  },
+};
   
-  module.exports = {
-    query: (text: string, params: string[], callback: Function) => {
-      console.log('executed query', text);
-      return pool.query(text, params, callback);
-    },
-  };
+export default db;
   
