@@ -1,12 +1,10 @@
 import React from 'react';
 
-import {useState, useEffect, useContext, createContext} from 'react'
-
 const App = () => {
 
-return (
+  return (
     <div>hello world</div>
-)
-}
+  );
+};
 
 export default App;
