@@ -1,6 +1,6 @@
-import React from 'react'
-import { createRoot } from 'react-dom/client'
-import App from './App'
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './App';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
